+++ test/queues/delay_test.js
@@ -5,7 +5,7 @@ const ironium = require('../../src');
 
 describe('queue with delay', ()=> {
 
-  const capture = ironium.queue('capture');
+  const captureQueue = ironium.queue('capture');
   // Allow up to 2s of delay when running this test suite.
   ironium.config.queues.maxDelay = 2000;
 
@@ -13,14 +13,14 @@ describe('queue with delay', ()=> {
   const processed = [];
 
   before(function() {
-    capture.each((job, callback)=> {
+    captureQueue.eachJob((job, callback)=> {
       processed.push(job);
       callback();
     });
   });
 
-  before(()=> capture.delay('delayed', '2s'));
-  before(ironium.once);
+  before(()=> captureQueue.delayJob('delayed', '2s'));
+  before(ironium.runOnce);
 
   it('should not process immediately', ()=>{
     assert.equal(processed.length, 0);
@@ -30,7 +30,7 @@ describe('queue with delay', ()=> {
     before(function(done) {
       setTimeout(done, 1500);
     });
-    before(ironium.once);
+    before(ironium.runOnce);
 
     it('should not process job', ()=>{
       assert.equal(processed.length, 0);
@@ -41,7 +41,7 @@ describe('queue with delay', ()=> {
     before(function(done) {
       setTimeout(done, 1000);
     });
-    before(ironium.once);
+    before(ironium.runOnce);
 
     it('should process job', ()=>{
       assert.equal(processed.length, 1);